Renamed CarAI to CarController, made AI, weapon/system panels and movement scripts components of it. Created a Panel base class, apply unlit material to panels.
