--- Assets/Scripts/Car/SystemsPanel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b0a2dee622781a94688f3d25d53f886e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
