--- Assets/Scripts/Car/WeaponsPanel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 416d260df88388442adfdfb08d65e7e0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
